refactor: extract Form_Records_List_Table from Form_Submissions
Move the Submissions list-table presentation (columns, sortable columns,
bulk/row-action definitions, per-column values, delivery badge, unread-row
styling) into a dedicated Form_Records_List_Table class. Status mutations
(bulk handler, untrash restore) stay in Form_Submissions. Add a
characterization test for the delivery column via the custom-column hook.
